CodeIgniter 4 - basic concepts - v 0.8 - bootstrap forms and backend
At this stage of our project we linked external bootstrap style file and correct errors in forms. As next part we reworked menu with bootstrap styling. Tab menu is used for logged in user, also other menu items are selectively available only for logged in user. In about main page is content separated in a row in two columns. In main part take place photo gallery for now (simple picture mosaic with jQuery).