CodeIgniter 4 - basic concepts - v 0.9 - pagination
In this stage of our project focused on demonstration of work with codeigniter framework, we introduced pagination into a news section and guesbook.  We also personalised styles of pagination with creating new app/views/pagers template for default pagination links. For better look, there has been used bootstrap styles. All implementation are commented and provided with links for further reference. In about page is new short news section that dislay only a few articles, for ordering records in a database and limiting display are query builder functionality in appropriate model (news model) used.