CodeIgniter 4 - basic concepts - v 0.9.3 - publishing/ unpublishing finished, some visual tweaks
In this stage of our project we fully implemented functionality for publishing and unpublishing news article on logged in admin page. Further visual improvements as different background image for common unlogegd user and for loged in user has been implemented. Unpublished articles are visual highlighted if admin is logged in. Also publish and unpublish buttons are deferentially colored for better visual marking.